requirements.txt
Django == 4.2.2
dropbox == 11.36.2
psycopg2==2.9.5
django-environ==0.9.0
requests==2.31.0
django-model-utils==4.3.1
black==23.3.0